+++ src/gpio.rs
@@ -73,7 +73,11 @@ mod hal_02;
 
 pub use embedded_hal::digital::v2::PinState;
 
+use core::fmt;
+
 /// A filler pin type
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct NoPin;
 
 /// Extension trait to split a GPIO peripheral in independent pins and registers
@@ -127,6 +131,7 @@ pub struct Analog;
 pub type Debugger = Alternate<PushPull, 0>;
 
 /// GPIO Pin speed selection
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Speed {
     Low = 0,
@@ -135,6 +140,7 @@ pub enum Speed {
     VeryHigh = 3,
 }
 
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Edge {
     Rising,
@@ -261,6 +267,24 @@ pub struct Pin<MODE, const P: char, const N: u8> {
 impl<MODE, const P: char, const N: u8> Pin<MODE, P, N> {
     const fn new() -> Self {
         Self { _mode: PhantomData }
+    }
+}
+
+impl<MODE, const P: char, const N: u8> fmt::Debug for Pin<MODE, P, N> {
+    fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_fmt(format_args!(
+            "P{}{}<{}>",
+            P,
+            N,
+            crate::stripped_type_name::<MODE>()
+        ))
+    }
+}
+
+#[cfg(feature = "defmt")]
+impl<MODE, const P: char, const N: u8> defmt::Format for Pin<MODE, P, N> {
+    fn format(&self, f: defmt::Formatter) {
+        defmt::write!(f, "P{}{}<{}>", P, N, crate::stripped_type_name::<MODE>());
     }
 }