+++ src/syscfg.rs
@@ -1,5 +1,6 @@
 use crate::pac::{RCC, SYSCFG};
 use crate::rcc::Enable;
+use core::fmt;
 use core::ops::Deref;
 
 /// Extension trait that constrains the `SYSCFG` peripheral
@@ -30,5 +31,18 @@ impl Deref for SysCfg {
     #[inline(always)]
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+#[cfg(feature = "defmt")]
+impl defmt::Format for SysCfg {
+    fn format(&self, f: defmt::Formatter) {
+        defmt::write!(f, "SysCfg(SYSCFG)");
+    }
+}
+
+impl fmt::Debug for SysCfg {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SysCfg").finish()
     }
 }