+++ CHANGELOG.md
@@ -26,7 +26,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - VSCode setting file
 - Add CAN1 PB8/PB9 and SPI3 MOSI PC1 pin mappings for F446 [#421]
 - Add embedded-storage traits for flash [#429]
+- `Debug` & `defmt::Format` impls for some structures and enums [#385]
 
+[#385]: https://github.com/stm32-rs/stm32f4xx-hal/pull/385
 [#421]: https://github.com/stm32-rs/stm32f4xx-hal/pull/421
 [#422]: https://github.com/stm32-rs/stm32f4xx-hal/pull/422
 [#423]: https://github.com/stm32-rs/stm32f4xx-hal/pull/423